refactor BiomeSource to pipeline addon
