+++ common/addons/biome-provider-pipeline/src/main/java/com/dfsek/terra/addons/biome/pipeline/config/BiomePipelineTemplate.java
@@ -8,7 +8,7 @@ import com.dfsek.terra.addons.biome.pipeline.api.Stage;
 import com.dfsek.terra.api.TerraPlugin;
 import com.dfsek.terra.api.config.meta.Meta;
 import com.dfsek.terra.api.world.biome.generation.BiomeProvider;
-import com.dfsek.terra.api.world.biome.generation.pipeline.BiomeSource;
+import com.dfsek.terra.addons.biome.pipeline.source.BiomeSource;
 
 import java.util.List;
