+++ common/addons/biome-provider-pipeline/src/main/java/com/dfsek/terra/addons/biome/pipeline/config/SourceTemplate.java
@@ -1,7 +1,7 @@
 package com.dfsek.terra.addons.biome.pipeline.config;
 
 import com.dfsek.tectonic.loading.object.ObjectTemplate;
-import com.dfsek.terra.api.world.biome.generation.pipeline.BiomeSource;
+import com.dfsek.terra.addons.biome.pipeline.source.BiomeSource;
 
 public abstract class SourceTemplate implements ObjectTemplate<BiomeSource>{
