+++ common/addons/biome-provider-pipeline/src/main/java/com/dfsek/terra/addons/biome/pipeline/source/NoiseSource.java
@@ -3,7 +3,6 @@ package com.dfsek.terra.addons.biome.pipeline.source;
 import com.dfsek.terra.api.noise.NoiseSampler;
 import com.dfsek.terra.api.util.collection.ProbabilityCollection;
 import com.dfsek.terra.api.world.biome.TerraBiome;
-import com.dfsek.terra.api.world.biome.generation.pipeline.BiomeSource;
 
 public class NoiseSource implements BiomeSource {
     private final ProbabilityCollection<TerraBiome> biomes;
